Create README.md
It is a simple Windows Forms application that allows you to create nodes of a directed graph, as well as add arcs. There is also saving the graph to a file and uploading the graph from a file (without saving the coordinates of the nodes). The program also implements the ability to find an arc with maximum multiplicity.

Also, through the use of double buffering, has been added the ability to move nodes with the mouse